+++ PowerShell/Profile/functions.ps1
@@ -2,6 +2,24 @@
 # PowerShell Core Profile Functions
 # ---------------------------------
 
+Function Update-ProfileModules {
+
+  $modpath = ($env:PSModulePath -split ";")[0]
+  $ymlpath = "$modpath\modules.yml"
+  $mods = (Get-ChildItem $modpath -Directory).Name
+  ConvertTo-Yaml -Data $mods -OutFile $ymlpath -Force
+
+  # Set-Location "$HOME\Documents"
+  # git pull
+  # git add PowerShell/Modules/**
+  # git commit -m "config: Updated modules configurations"
+  # git-cliff -o "$HOME\Documents\CHANGELOG.md"
+  # git add CHANGELOG.md
+  # git commit -m "doc: update CHANGELOG.md for added modules"
+  # git push
+
+}
+
 # ---------------------
 # Search via `s-cli`
 # ---------------------